clean up of parsing
