+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: abuzdin <abuzdin@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 18:11:16 by mthiry            #+#    #+#             */
-/*   Updated: 2022/11/25 20:02:23 by abuzdin          ###   ########.fr       */
+/*   Updated: 2022/11/25 21:57:47 by abuzdin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -277,6 +277,7 @@ char	*cub_strdup(const char *s, t_data *data);
 int		ft_strcmp(char *s1, char *s2);
 char	*get_next_line(int fd, t_data *data);
 int		check_charset(char c, char *charset);
+int		ft_atoi_er(const char *str, int *error);
 
 /* cub3d.c */
 void	hook_manager(t_data *data);
@@ -346,8 +347,6 @@ int		load_textures(t_data *data);
 
 /* init_map_utils_1.c */
 void	check_chars(char **raw, t_data *data);
-char	*find_param(char **raw, char *param, t_data *data);
-char	*find_param_color(char **raw, char *param, t_data *data);
 size_t	check_lines(char **map, size_t j, t_data *data);
 void	find_mapsize(char **raw, int j, t_data *data);
 void	copy_map(char **raw, t_data *data);
@@ -363,8 +362,14 @@ void	check_columns(char **map, t_data *data);
 /* init_map_utils_3.c */
 void	check_player(char **map, t_data *data);
 void	check_direction(t_data *data);
+void	check_num(char **str, int *array, t_data *data);
+void	check_colors(t_map *map, t_data *data);
 void	print_map(t_data *data, char **map);
-int		ft_atoi_er(const char *str, int *error);
+
+/* init_map_utils_4.c */
+char	*find_param(char **raw, char *param, t_data *data);
+char	*find_param_color(char **raw, char *param, t_data *data);
+
 
 /* init_map.c */
 void	init_vars(t_data *data);
