+++ init_map_utils_1.c
@@ -6,7 +6,7 @@
 /*   By: abuzdin <abuzdin@student.s19.be>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 18:09:01 by abuzdin           #+#    #+#             */
-/*   Updated: 2022/11/25 21:26:46 by abuzdin          ###   ########.fr       */
+/*   Updated: 2022/11/25 21:58:12 by abuzdin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,54 +34,6 @@ void	check_chars(char **raw, t_data *data)
 		}
 		++j;
 	}
-}
-
-// find a wall texture
-char	*find_param(char **raw, char *param, t_data *data)
-{
-	size_t	i;
-	size_t	len;
-	char	**tmp;
-	char	*ret;
-
-	i = 0;
-	len = ft_strlen(param);
-	while (raw[i] && ft_strncmp(raw[i], param, len))
-		++i;
-	if (!raw[i])
-		return (NULL);
-	if (i > data->j)
-		data->j = i;
-	tmp = ft_split(raw[i], ' ');
-	alloc_check_big(tmp, data);
-	ret = cub_strdup(tmp[1], data);
-	cub_free(tmp);
-	return (ret);
-}
-
-// find color setting for floor/ceiling
-char	*find_param_color(char **raw, char *param, t_data *data)
-{
-	size_t	i;
-	size_t	j;
-	size_t	len;
-	char	*ret;
-
-	j = 0;
-	len = ft_strlen(param);
-	while (raw[j] && ft_strncmp(raw[j], param, len))
-		++j;
-	if (!raw[j])
-		return (NULL);
-	if (j > data->j)
-		data->j = j;
-	i = 0;
-	while (raw[j][i] && !ft_isdigit(raw[j][i]))
-		++i;
-	if (!raw[j][i])
-		return (NULL);
-	ret = cub_strdup(raw[j] + i, data);
-	return (ret);
 }
 
 // count lines in the map
